# [CH] optimize master_commit_red_jobs by eliminating join and adding pre-filtering