@huydhn 2025-04-22
Is this a correct understand that in order to implement this optimization we need to update the schema of `workflow_job` table to add these new column?  I think we need to keep track of such changes eventually.  @yangw-dev and I have try to keep the schema of some tables on https://github.com/pytorch/test-infra/tree/main/clickhouse_db_schema.  It's not ideal because we still need to create/alter them manually, but we rarely need to do so.